fix: Inherit owner from parent folder (#951)

+++ src/install.sh
@@ -366,7 +366,9 @@ if [[ "${BOOT,,}" != "http"* ]]; then
   error "Invalid BOOT value specified, \"$BOOT\" is not a valid URL!" && exit 64
 fi
 
-mkdir -p "$STORAGE"
+if ! makeDir "$STORAGE"; then
+  error "Failed to create directory \"$STORAGE\" !" && exit 33
+fi
 
 find "$STORAGE" -maxdepth 1 -type f \( -iname '*.rom' -or -iname '*.vars' \) -delete
 find "$STORAGE" -maxdepth 1 -type f \( -iname 'data.*' -or -iname 'qemu.*' \) -delete
@@ -410,7 +412,11 @@ case "${base,,}" in
 
     tmp="$STORAGE/extract"
     rm -rf "$tmp"
-    mkdir -p "$tmp"
+
+    if ! makeDir "$tmp"; then
+      error "Failed to create directory \"$tmp\" !" && exit 33
+    fi
+
     7z x "$STORAGE/$base" -o"$tmp" > /dev/null
 
     rm -f "$STORAGE/$base"
@@ -429,6 +435,8 @@ esac
 
 case "${base,,}" in
   *".iso" | *".img" | *".raw" | *".qcow2" )
+
+    ! setOwner "$STORAGE/$base" && error "Failed to set the owner for \"$STORAGE/$base\" !"
     detectType "$STORAGE/$base" && return 0
     error "Cannot read file \"${base}\"" && exit 63 ;;
 esac
@@ -451,6 +459,8 @@ dst="$STORAGE/${base%.*}.$target_ext"
 ! convertImage "$STORAGE/$base" "$source_fmt" "$dst" "$target_fmt" && exit 35
 
 base=$(basename "$dst")
+
+! setOwner "$STORAGE/$base" && error "Failed to set the owner for \"$STORAGE/$base\" !"
 detectType "$STORAGE/$base" && return 0
 error "Cannot convert file \"${base}\"" && exit 36
